fix int casting

+++ functions/benchmarks/causal_bench_1M.py
@@ -153,7 +153,7 @@ def run(flconn, kvs, mode, segment, params):
         ### CREATE DATA###
         logging.info('Warming up keys')
         warm_begin = time.time()
-        block_size = total_num_keys/6
+        block_size = int(total_num_keys/6)
         for k in range(block_size*segment+1,block_size*segment + block_size+1):
             if k % 50 == 0:
                 logging.info('warmup for key %s done' % k)
